% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nanoflann_nnIndex}
\alias{nanoflann_nnIndex}
\alias{nnIndex}
\title{Eigen Matrix nanoflann nearest neighbor indexes}
\usage{
nnIndex(P, Q)
}
\arguments{
\item{P}{dataframe (columns 1,2 being x,y) of data points.}

\item{Q}{dataframe (columns 1,2 being x,y) of query points.}
}
\value{
vector of descending unqiue nearest neighbor indexes
}
\description{

}

