% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcppnanoflann.R
\docType{Rcppnanoflann}
\name{Rcppnanoflann}
\alias{Rcppnanoflann}
\title{stackoverflow answer}
\description{
For question
[Closest point to a path](http://stackoverflow.com/questions/27321856/closest-point-to-a-path)
}
\concept{
This answer illustrates how the nanoflann library can
be used along with Rcpp, RcppEigen, and C++11 to easily create an
installable package answering the original question.
}

